OpenAssistant/oasst-sft-4-pythia-12b-epoch-3.5
