+++ .github/CODEOWNERS
@@ -1,0 +1,2 @@
+* @JV-conseil
+
