+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,70 @@
+---
+name: Bug Report 🐞
+about: Did you find a bug?
+title: 'Bug Report 🐞 {short-description}'
+labels: bug, triage
+assignees: 'JV-conseil'
+
+---
+
+<!--
+
+  Hi there 👋 Thank you for discovering and submitting an issue.
+
+  Before you submit this; let's make sure of a few things.
+  Please make sure the following boxes are ticked if they are correct.
+  If not, please try and fulfill these first.
+
+-->
+
+- [ ] I have activated the browser Private mode to evaluate if the error still happens.
+- [ ] I have turn-off any extension on my browser like adblock that may interfere with the app behaviour.
+- [ ] I am on the [latest][latest] version.
+- [ ] I have checked that no other similar [issues][issues] or [discussions][discussions] are already opened and believe that this is not a duplicate.
+- [ ] If an exception occurs when executing a command, I execute it again in debug mode (`DEBUF = True` in settings).
+
+---
+
+### Describe the bug 🐛
+
+> A clear and precise description of what the bug is, please be descriptive! Thanks 🙌
+
+### To Reproduce 🔂
+
+> Steps to reproduce the behavior:
+>
+> 1. Go to '...'
+> 2. Click on '....'
+> 3. Scroll down to '....'
+> 4. See error
+
+### Expected behavior 🚀
+
+> A clear and concise description of what you expected to happen.
+
+### Configuration ⚙️
+
+> A clear and concise description of your configuration.
+>
+> Python 3.11.0
+> PostgreSQL Server 14.5
+> Django version 4.1.2
+> ...
+
+### Screenshots 📸
+
+> If applicable, add screenshots to help explain your problem.
+
+### Diagnosis attempts 🩺
+
+> `curl "https://some.domain.name" --verbose`
+
+### Additional context 🌍
+
+> Add any other context about the problem here.
+
+<!-- links -->
+
+[discussions]: https://github.com/ulynks/www/discussions
+[issues]: https://github.com/ulynks/www/issues
+[latest]: https://github.com/ulynks/www/releases/latest
