+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,36 @@
+---
+name: Feature Request 🎁
+about: Do you have ideas for new features and improvements?
+title: ''
+labels: feature, triage
+assignees: ''
+
+---
+
+<!--
+  Hi there! Thank you for discovering and submitting an issue.
+
+  Before you submit this; let's make sure of a few things.
+  Please make sure the following boxes are ticked if they are correct.
+  If not, please try and fulfill these first.
+-->
+
+<!-- Checked checkbox should look like this: [x] -->
+- [ ] I have searched the [issues] of this repo and believe that this is not a duplicate.
+- [ ] If an exception occurs when executing a command, I executed it again in debug mode (`DEBUF = True` in settings).
+
+**Is your feature request related to a problem? Please describe.**
+A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
+
+**Describe the solution you'd like**
+A clear and concise description of what you want to happen.
+
+**Describe alternatives you've considered**
+A clear and concise description of any alternative solutions or features you've considered.
+
+**Additional context**
+Add any other context or screenshots about the feature request here.
+
+<!-- Definition -->
+
+[issues]: https://github.com/ulynks/www/issues
