+++ app/api/chat/route.tsx
@@ -1,29 +1,34 @@
 /** @jsxImportSource ai-jsx */
-import { toTextStream } from 'ai-jsx/stream'
-import {
-  ChatCompletion,
-  ConversationHistory,
-  SystemMessage
-} from 'ai-jsx/core/completion'
-import { Prompt } from 'ai-jsx/batteries/prompts'
-import { StreamingTextResponse } from 'ai'
-import { PropsOfComponent } from 'ai-jsx'
+import { HfInference } from '@huggingface/inference';
+import { HuggingFaceStream, StreamingTextResponse } from 'ai';
+import { experimental_buildOpenAssistantPrompt } from 'ai/prompts';
 
-function App({ messages }: { messages: PropsOfComponent<typeof ConversationHistory>['messages'] }) {
-  return (
-    <ChatCompletion temperature={1}>
-      <SystemMessage>
-        <Prompt hhh persona="expert translator" />
-        Translate any messages you receive to French. Respond only with the
-        translation.
-      </SystemMessage>
-      <ConversationHistory messages={messages} />
-    </ChatCompletion>
-  )
-}
+// Create a new HuggingFace Inference instance
+const Hf = new HfInference(process.env.HUGGINGFACE_API_KEY);
+
+// IMPORTANT! Set the runtime to edge
+export const runtime = 'edge';
 
 export async function POST(req: Request) {
-  const { messages } = await req.json()
+  // Extract the `messages` from the body of the request
+  const { messages } = await req.json();
 
-  return new StreamingTextResponse(toTextStream(<App messages={messages} />))
+  const response = Hf.textGenerationStream({
+    model: 'OpenAssistant/oasst-sft-4-pythia-12b-epoch-3.5',
+    inputs: experimental_buildOpenAssistantPrompt(messages),
+    parameters: {
+      max_new_tokens: 200,
+      // @ts-ignore (this is a valid parameter specifically in OpenAssistant models)
+      typical_p: 0.2,
+      repetition_penalty: 1,
+      truncate: 1000,
+      return_full_text: false,
+    },
+  });
+
+  // Convert the response into a friendly text-stream
+  const stream = HuggingFaceStream(response);
+
+  // Respond with the stream
+  return new StreamingTextResponse(stream);
 }
