+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "next-template",
+  "name": "fixie-vercel-ai-chatbot-ai-jsx",
   "version": "0.0.2",
   "private": true,
   "scripts": {
@@ -14,23 +14,24 @@
     "format:check": "prettier --check \"{app,lib,components}**/*.{ts,tsx,mdx}\" --cache"
   },
   "dependencies": {
-    "@radix-ui/react-alert-dialog": "^1.0.4",
-    "@radix-ui/react-dialog": "^1.0.4",
-    "@radix-ui/react-dropdown-menu": "^2.0.5",
+    "@huggingface/inference": "^2.6.4",
+    "@radix-ui/react-alert-dialog": "^1.0.5",
+    "@radix-ui/react-dialog": "^1.0.5",
+    "@radix-ui/react-dropdown-menu": "^2.0.6",
     "@radix-ui/react-label": "^2.0.2",
     "@radix-ui/react-select": "^1.2.2",
     "@radix-ui/react-separator": "^1.0.3",
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-switch": "^1.0.3",
-    "@radix-ui/react-tooltip": "^1.0.6",
-    "@vercel/analytics": "^1.0.0",
-    "@vercel/kv": "^0.2.1",
-    "@vercel/og": "^0.5.7",
-    "ai": "^2.1.6",
-    "ai-jsx": "^0.5.11",
+    "@radix-ui/react-tooltip": "^1.0.7",
+    "@vercel/analytics": "^1.2.0",
+    "@vercel/kv": "^0.2.4",
+    "@vercel/og": "^0.5.20",
+    "ai": "^2.2.35",
+    "ai-jsx": "^0.5.16",
     "class-variance-authority": "^0.4.0",
     "clsx": "^1.2.1",
-    "focus-trap-react": "^10.1.1",
+    "focus-trap-react": "^10.2.3",
     "nanoid": "^4.0.2",
     "next": "13.4.7-canary.1",
     "next-auth": "0.0.0-manual.83c4ebd1",
@@ -39,31 +40,31 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hot-toast": "^2.4.1",
-    "react-intersection-observer": "^9.4.4",
+    "react-intersection-observer": "^9.8.0",
     "react-markdown": "^8.0.7",
     "react-syntax-highlighter": "^15.5.0",
-    "react-textarea-autosize": "^8.4.1",
+    "react-textarea-autosize": "^8.5.3",
     "remark-gfm": "^3.0.1",
     "remark-math": "^5.1.1"
   },
   "devDependencies": {
-    "@tailwindcss/typography": "^0.5.9",
-    "@types/node": "^17.0.12",
-    "@types/react": "^18.0.22",
-    "@types/react-dom": "^18.0.7",
-    "@types/react-syntax-highlighter": "^15.5.6",
-    "@typescript-eslint/parser": "^5.59.7",
-    "autoprefixer": "^10.4.13",
-    "eslint": "^8.31.0",
+    "@tailwindcss/typography": "^0.5.10",
+    "@types/node": "^17.0.45",
+    "@types/react": "^18.2.56",
+    "@types/react-dom": "^18.2.19",
+    "@types/react-syntax-highlighter": "^15.5.11",
+    "@typescript-eslint/parser": "^5.62.0",
+    "autoprefixer": "^10.4.17",
+    "eslint": "^8.56.0",
     "eslint-config-next": "13.4.7-canary.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-tailwindcss": "^3.12.0",
-    "postcss": "^8.4.21",
-    "prettier": "^2.7.1",
-    "tailwind-merge": "^1.12.0",
-    "tailwindcss": "^3.3.1",
-    "tailwindcss-animate": "^1.0.5",
-    "typescript": "^5.1.3"
+    "eslint-config-prettier": "^8.10.0",
+    "eslint-plugin-tailwindcss": "^3.14.2",
+    "postcss": "^8.4.35",
+    "prettier": "^2.8.8",
+    "tailwind-merge": "^1.14.0",
+    "tailwindcss": "^3.4.1",
+    "tailwindcss-animate": "^1.0.7",
+    "typescript": "^5.3.3"
   },
   "packageManager": "pnpm@8.6.3"
 }
